feat: Ajoute les caractéristiques associées au produit (non fonctionel)

+++ src/main/java/com/projet/ecommerce/business/impl/CaracteristiqueBusiness.java
@@ -5,7 +5,6 @@ import com.projet.ecommerce.business.dto.CaracteristiqueDTO;
 import com.projet.ecommerce.business.dto.transformer.CaracteristiqueTransformer;
 import com.projet.ecommerce.entrypoint.graphql.GraphQLCustomException;
 import com.projet.ecommerce.persistance.entity.Caracteristique;
-import com.projet.ecommerce.persistance.entity.Produit;
 import com.projet.ecommerce.persistance.repository.CaracteristiqueRepository;
 import com.projet.ecommerce.persistance.repository.CaracteristiqueRepositoryCustom;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -47,24 +46,20 @@ public class CaracteristiqueBusiness implements ICaracteristiqueBusiness {
             if (caracteristiqueOptional.isPresent()) {
                 throw new GraphQLCustomException("Erreur dans l'ajout de la caractéristique : La caractéristique avec ce label existe déjà.");
             }
-
         }*/
 
         // Si pas d'erreur
         Caracteristique caracteristique = CaracteristiqueTransformer.dtoToEntity(caracteristiqueDTO);
-        caracteristiqueRepository.save(caracteristique);
-
-        return caracteristiqueDTO;
-        // TODO à remplacer par la ligne ci-dessous losque les fonctions get auront été implémentées
-        // return CaracteristiqueTransformer.entityToDto(caracteristiqueRepository.save(caracteristique));
+        return CaracteristiqueTransformer.entityToDto(caracteristiqueRepository.save(caracteristique));
     }
 
     @Override
     public boolean delete(int idCaracteristique) {
 
         Optional<Caracteristique> caracteristiqueOptional = caracteristiqueRepository.findById(idCaracteristique);
-        //caracteristiqueOptional.orElseThrow(() -> {throw new GraphQLCustomException("Aucune caractéristique attaché à cet id.");});
-
+        if (!caracteristiqueOptional.isPresent()) {
+            throw new GraphQLCustomException("Aucune caractéristique attaché à cet id.");
+        }
         try {
             caracteristiqueRepository.deleteById(idCaracteristique);
             return true;
